refactor signup and use reusable signform Relates #84
