+++ client/src/pages/signup/index.jsx
@@ -1,13 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { GoogleOutlined } from '@ant-design/icons';
 
 import axios from 'axios';
 import {
-  Form,
   Button,
-  Input,
   message,
   Space,
   Typography,
@@ -16,21 +14,17 @@ import {
 
 import { sign } from '../../redux/feature/user/userSlice';
 import './style.less';
+import SignForm from '../../components/common/SignForm';
 
 const { Link } = Anchor;
 
-const { Password } = Input;
-const { Item } = Form;
 const { Title, Text } = Typography;
 
 function SignUp() {
   const navigate = useNavigate();
   const dispatch = useDispatch();
-  const [userInfo, setUserInfo] = useState({
-    name: '', email: '', password: '', phone: '', address: 'Gaza',
-  });
 
-  const registor = async () => {
+  const register = async (userInfo) => {
     try {
       const { data: { data } } = await axios.post('/api/signup', userInfo);
       dispatch(sign(data));
@@ -42,26 +36,7 @@ function SignUp() {
       });
     }
   };
-  const passwordValidation = () => ({
-    validator(_, value) {
-      if (value.length >= 6) {
-        return Promise.resolve();
-      }
-      return Promise.reject(new Error('password should have at least 6 character '));
-    },
-  });
-  const confirmPasswordValidation = (getFieldValue) => ({
-    validator(_, value) {
-      if (!value || getFieldValue('password') === value) {
-        return Promise.resolve();
-      }
-      return Promise.reject(new Error('The two passwords that you entered do not match!'));
-    },
-  });
 
-  const handleChange = ({ target: { name, value } }) => {
-    setUserInfo({ ...userInfo, [name]: value });
-  };
   return (
     <div className="sign-up-container">
       <div className="img-side-sgin-up">
@@ -83,102 +58,16 @@ function SignUp() {
           >
             REGISTER
           </Title>
-          <Form
-            className="Form-sign-up"
-            name="register"
-            onFinish={registor}
-          >
-            <Item
-              name="name"
-              rules={[
-                {
-                  required: true,
-                  message: 'Please input your name!',
-                  whitespace: true,
-                },
-              ]}
-            >
-              <Input
-                name="name"
-                placeholder="Name"
-                onChange={(e) => handleChange(e)}
+          <SignForm getUserInfo={register} type="register" />
+          <div className="register_option">
+            <Text>Already have an account ?</Text>
+            <Anchor affix={false}>
+              <Link
+                href="/login"
+                title="Sign in"
               />
-            </Item>
-            <Item
-              name="email"
-              rules={[
-                {
-                  type: 'email',
-                  message: 'The input is not valid E-mail!',
-                },
-                {
-                  required: true,
-                  message: 'Please input your E-mail!',
-                },
-              ]}
-            >
-              <Input
-                name="email"
-                placeholder="Email"
-                onChange={(e) => handleChange(e)}
-              />
-            </Item>
-            <Item
-              name="phone"
-              rules={[
-                {
-                  required: true,
-                  message: 'Please input your phone number!',
-                },
-              ]}
-            >
-              <Input
-                placeholder="Phone Number"
-                name="phone"
-                onChange={(e) => handleChange(e)}
-              />
-            </Item>
-            <Item
-              name="password"
-              rules={[
-                {
-                  required: true,
-                  message: 'Please input your password!',
-                },
-                () => passwordValidation(),
-              ]}
-            >
-              <Password
-                placeholder="Password"
-                name="password"
-                onChange={(e) => handleChange(e)}
-              />
-            </Item>
-            <Item
-              name="confirm"
-              rules={[
-                {
-                  required: true,
-                  message: 'Please confirm your password!',
-                },
-                ({ getFieldValue }) => confirmPasswordValidation(getFieldValue),
-              ]}
-            >
-              <Password placeholder="Confirm password" />
-            </Item>
-            <Button className="sign-up-btn" type="primary" htmlType="submit">
-              Sign Up
-            </Button>
-            <div className="register_option">
-              <Text>Already have an account ?</Text>
-              <Anchor affix={false}>
-                <Link
-                  href="/login"
-                  title="Sign in"
-                />
-              </Anchor>
-            </div>
-          </Form>
+            </Anchor>
+          </div>
           <Button type="primary" icon={<GoogleOutlined />}>
             Sign in with Google
           </Button>
